Rearrange bottom panel related sources
